Completed - work on bugs v2.2
